Update location for SetupPostgresDB
